requirements.txt
Django==1.3.1
PIL==1.1.7
django-debug-toolbar==0.8.5
-e git://github.com/dcramer/django-paypal.git@7ae72d86b01d57b76f3aa9b8001ff6863988c867#egg=django_paypal-0.1.2-py2.6-dev
django-templatetag-sugar==0.1
sorl-thumbnail==11.09
